+++ src/walkgame/objects/parentObjects/Enemy.java
@@ -5,7 +5,7 @@ import javafx.scene.image.ImageView;
 
 public class Enemy extends Character
 {
-    public Enemy(double x, double y, Image image, String name, int health, float speed)
+    public Enemy(double x, double y, Image image, String name, int health, double speed)
     {
         super(x, y, image, name, health, speed);
     }
